Moved class SimulatorNode into its own h-file.
